+++ packages/chat/src/chat-bubble/bubble-info.tsx
