+++ packages/chat/src/chat-bubble/chat-bubble-ui.stories.tsx
@@ -1,9 +1,11 @@
 import React from 'react'
 import { Story } from '@storybook/react'
 
-import { MessageType } from './types'
-import { MEDIA_ARGS, SAMPLE_IMAGES } from './bubbles/__stories__/constants'
+import { MEDIA_ARGS, SAMPLE_IMAGES } from '../bubbles/__stories__/constants'
+
 import { ChatBubbleUI, ChatBubbleUIProps } from './chat-bubble-ui'
+
+import { MessageType } from '@titicaca/chat'
 
 export default {
   title: 'chat-bubble / ChatBubbleUI',
