+++ packages/chat/src/chat-bubble/index.ts
@@ -1,0 +1,1 @@
+export { default } from './chat-bubble'
