+++ packages/chat/src/index.tsx
@@ -1,3 +1,3 @@
 export * from './types'
 export * from './bubbles'
-export { default } from './chat-bubble'
+export * as ChatBubble from './chat-bubble'
